# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavecrest Pro API",
    "version": "1.0.0",
    "description": "Wavecrest Pro is a real-time trend intelligence platform that aggregates trending topics across multiple social platforms. This API provides endpoints for authentication, trend discovery, dashboard management, billing, and analytics.",
    "contact": {
      "name": "Wavecrest Pro Support",
      "url": "https://wavecrest.pro",
      "email": "support@wavecrest.pro"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://wavecrest.pro",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/trends/search": {
      "get": {
        "tags": ["Trends"],
        "summary": "Search trends by keyword",
        "description": "Full-text search across all trending topics. Requires Pro plan or higher.",
        "operationId": "searchTrends",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search query string",
            "schema": { "type": "string" }
          },
          {
            "name": "platform",
            "in": "query",
            "description": "Filter results by platform",
            "schema": { "type": "string", "enum": ["twitter", "reddit", "tiktok", "youtube", "google", "github"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                },
                "example": {
                  "query": "AI",
                  "platform": null,
                  "count": 12,
                  "results": [
                    {
                      "topic": "OpenAI GPT-5 Launch",
                      "score": 97,
                      "platform": "twitter",
                      "fetched_at": "2025-01-15T14:30:00.000Z"
                    },
                    {
                      "topic": "AI Video Generation",
                      "score": 85,
                      "platform": "youtube",
                      "fetched_at": "2025-01-15T12:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing search query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Query parameter 'q' is required"
                }
              }
            }
          },
          "403": {
            "description": "Plan upgrade required (Pro+ only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Search requires a Pro plan or higher"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analytics/velocity": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get trend velocity data",
        "description": "Returns the number of new trends detected per day over a given period, along with the daily average. Requires Pro plan or higher.",
        "operationId": "getVelocity",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Number of days to analyze (default 7, max 90)",
            "schema": { "type": "integer", "default": 7, "maximum": 90 }
          }
        ],
        "responses": {
          "200": {
            "description": "Velocity data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VelocityResponse"
                },
                "example": {
                  "days": 7,
                  "dates": ["2025-01-09", "2025-01-10", "2025-01-11", "2025-01-12", "2025-01-13", "2025-01-14", "2025-01-15"],
                  "counts": [180, 210, 195, 220, 248, 232, 248],
                  "avg": 219
                }
              }
            }
          },
          "403": {
            "description": "Plan upgrade required (Pro+ only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Analytics requires a Pro plan or higher"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/platforms": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get platform breakdown",
        "description": "Returns the distribution of trending topics across platforms for the current day. Requires Pro plan or higher.",
        "operationId": "getPlatformAnalytics",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Platform distribution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformAnalyticsResponse"
                },
                "example": {
                  "date": "2025-01-15",
                  "total": 248,
                  "platforms": [
                    { "platform": "twitter", "count": 62, "percentage": 25.0 },
                    { "platform": "reddit", "count": 48, "percentage": 19.4 },
                    { "platform": "tiktok", "count": 45, "percentage": 18.1 },
                    { "platform": "youtube", "count": 38, "percentage": 15.3 },
                    { "platform": "google", "count": 35, "percentage": 14.1 },
                    { "platform": "github", "count": 20, "percentage": 8.1 }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Plan upgrade required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/top-trends": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get top trending topics",
        "description": "Returns the highest-scoring trending topics across all platforms, optionally filtered by time range. Requires Pro plan or higher.",
        "operationId": "getTopTrends",
        "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "description": "Look-back period in days (default 1, max 30)",
            "schema": { "type": "integer", "default": 1, "maximum": 30 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of top trends to return (default 10, max 50)",
            "schema": { "type": "integer", "default": 10, "maximum": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "Top trending topics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopTrendsResponse"
                },
                "example": {
                  "days": 1,
                  "count": 10,
                  "trends": [
                    { "topic": "OpenAI GPT-5 Launch", "score": 97, "platform": "twitter", "category": "technology" },
                    { "topic": "Apple Vision Pro 2", "score": 94, "platform": "twitter", "category": "technology" },
                    { "topic": "Super Bowl LIX", "score": 91, "platform": "google", "category": "sports" },
                    { "topic": "Bitcoin ATH", "score": 89, "platform": "reddit", "category": "finance" },
                    { "topic": "GTA VI Trailer 2", "score": 88, "platform": "youtube", "category": "gaming" }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Plan upgrade required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/stats": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get aggregate statistics",
        "description": "Returns high-level statistics about the trend database: total trends tracked, platforms active, average daily trends, and more. Requires Pro plan or higher.",
        "operationId": "getStats",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Aggregate statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                },
                "example": {
                  "total_trends": 18420,
                  "total_platforms": 6,
                  "avg_daily_trends": 219,
                  "avg_score": 64.3,
                  "top_platform": "twitter",
                  "top_category": "technology",
                  "data_since": "2024-06-01T00:00:00.000Z",
                  "last_updated": "2025-01-15T14:30:00.000Z"
                }
              }
            }
          },
          "403": {
            "description": "Plan upgrade required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie set after login or signup"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "properties": {
          "query": { "type": "string" },
          "platform": { "type": "string", "nullable": true },
          "count": { "type": "integer" },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "topic": { "type": "string" },
                "score": { "type": "integer" },
                "platform": { "type": "string" },
                "fetched_at": { "type": "string", "format": "date-time" }
              }
            }
          }
        }
      },
...
      "VelocityResponse": {
        "type": "object",
        "properties": {
          "days": { "type": "integer" },
          "dates": { "type": "array", "items": { "type": "string", "format": "date" } },
          "counts": { "type": "array", "items": { "type": "integer" } },
          "avg": { "type": "number" }
        }
      },
      "PlatformAnalyticsResponse": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "total": { "type": "integer" },
          "platforms": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "platform": { "type": "string" },
                "count": { "type": "integer" },
                "percentage": { "type": "number" }
              }
            }
          }
        }
      },
      "TopTrendsResponse": {
        "type": "object",
        "properties": {
          "days": { "type": "integer" },
          "count": { "type": "integer" },
          "trends": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "topic": { "type": "string" },
                "score": { "type": "integer" },
                "platform": { "type": "string" },
                "category": { "type": "string" }
              }
            }
          }
        }
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "total_trends": { "type": "integer" },
          "total_platforms": { "type": "integer" },
          "avg_daily_trends": { "type": "number" },
          "avg_score": { "type": "number" },
          "top_platform": { "type": "string" },
          "top_category": { "type": "string" },
          "data_since": { "type": "string", "format": "date-time" },
          "last_updated": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}